+++ src/components/note/note.js
@@ -42,7 +42,8 @@ Note.propTypes = {
     'legacy',
     'beta',
     'default',
-    'new'
+    'new',
+    'pricing'
   ])
 };