+++ src/components/themes/themes.js
@@ -100,6 +100,16 @@ const themes = {
       color: '#cf1c61',
       borderColor: '#fd8ac0'
     }
+  },
+  pricing: {
+    image: <Image icon="creditcard" color="green" />,
+    label: 'Pricing',
+    tooltipText: 'This contains information about product pricing.',
+    styles: {
+      background: '#e8f5ee',
+      color: '#15603d',
+      borderColor: '#378645'
+    }
   }
 };
 /* duplicate themes */